requirements.txt
openai==1.13.3
langchain==0.1.11
langchain-openai==0.0.8
python-dotenv
langchainhub==0.1.15
pandas==2.2.2
langchainhub